==> notebooks/workshop-1/01-data-preprocessing.ipynb <==
# Data preprocessing

Natural Language Processing (NLP) is a rapidly growing field that has numerous applications in various domains, including healthcare. The medical field has an abundance of data that can be utilized for NLP tasks, such as medical transcriptions and clinical notes. However, this data often requires extensive preprocessing before it can be used effectively for machine learning models.

The dataset we'll be using comes from the [socd6/medical-nlp](https://github.com/socd06/medical-nlp0) repository.
This dataset, compiled for NLP experiments, contains medical transcriptions and custom-generated clinical stop words and vocabulary. It includes a fully processed dataset simplified to 4 classes: Surgery, Medical Records, Internal Medicine, and Other. The dataset is divided into training and test subsets to allow for the development and evaluation of machine learning models. The dataset also contains the original medical specialties for each transcription.

## Objective

The objective of this workshop is to finetune a BERT (Bidirectional Encoder Representations from Transformers) model to classify medical transcriptions. BERT is a powerful NLP model that has achieved state-of-the-art results on a wide range of NLP tasks. By using BERT, we aim to improve the accuracy and effectiveness of our classification model.

In this Jupyter Notebook, we will explore the preprocessing steps required to prepare the dataset for use in the BERT model. We will generate two datasets:

 1. Fine-tuning dataset to have the BERT model better suit handling medical transcriptions.
 2. Validation dataset to inspect the classification accuracy of the BERT models.



## Instructions

The notebook has been implemented to require the user to write their own code. Normal descriptive information is in plain text.

> Any text within a quote block indicates instructions for the user.

## Initial load

Let's first begin by taking a look at our data.



> Use the pandas library to read in the medical-nlp dataset from the file and assign it to a variable named ``df``:
> - Local: `"../../data/medical-nlp/mtsamples.csv"`
> - Colab: `"./data/medical-nlp/mtsamples.csv"`
>
> Print out a list of the columns in the dataset.
>
> Display the first few rows of the dataset.

import pandas as pd
import seaborn as sns

RAW_DATA_PATH = "../../data/medical-nlp/mtsamples.csv"

sns.set()

df = pd.read_csv(RAW_DATA_PATH, index_col=0)
print("Columns:")
for c in df.columns:
    print(f"  {c}")
df.head()

Let's also review column-wise contents.


> Generate a summary of descriptive statistics for the numeric columns in the dataset
>
> Display the resulting summary statistics.

df.describe()

## Missing data handling

It seems that not every row have all the values. Let's investigate.

>Check the dataset for missing values.
>
>Display the number of missing values in each column of the dataset.

df.isnull().sum(axis=0)

Medical appointment transcriptions are full length texts that we are really interested in. Let's drop any rows that are missing the values for `transcription`.

>Remove any rows from the ``df`` dataset where the ``transcription`` column contains missing values.
>
>Display a summary of descriptive statistics for the columns in the resulting dataset (no numeric columns should be present).

df.dropna(axis="index", subset=["transcription"], inplace=True)
df.describe()

# Don't remove this cell
assert df.shape == (4966, 5)

## Exploration

In this section, we will explore the ``medical_specialty`` and ``transcription`` columns of the medical-nlp dataset in more detail. These columns contain valuable information about the medical specialties and corresponding transcriptions of the medical reports.

The ``medical_specialty`` column indicates the medical specialty associated with each report, such as cardiology, neurology, or orthopedics. This column may provide useful insights into the distribution of medical reports across different specialties and the potential differences in language use between them.

The ``transcription`` column contains the text of the medical reports themselves. This column is of particular interest as it contains the primary data we will be using to train and evaluate our machine learning models. The text data in this column will require extensive preprocessing to ensure that it is in a suitable format for use with our models.

In the following cells, we will explore the distribution of medical specialties in the dataset and examine some example transcriptions to gain a better understanding of the text data we are working with. We will also discuss the preprocessing steps required to prepare the text data for use with machine learning models.

### `medical_specialty`

Let's begin with the `medical_specialty` column. First we find out the unique values in the column.

>Retrieve the unique values from the ``medical_specialty`` column of the ``df`` dataset.
>
>Print out the number of unique medical specialties in the dataset.
>
>Display a list of the unique medical specialties.

medical_specialties = df["medical_specialty"].unique()
print(f"Unique medical specialties: {len(medical_specialties)}")
medical_specialties

Let's then visualize the distribution of medical specialties in our dataset.

>Count the number of occurrences of each medical specialty in the ``medical_specialty`` column of the ``df`` dataset.
>
>Generate a horizontal bar chart to visualize the distribution of medical specialties.
>
>Set the size of the chart to fit the image within the Jupyter Notebook cell and adjust the font size as needed.

df["medical_specialty"].value_counts().plot(kind="barh", figsize=(8, 10), fontsize=8)

From this we can see that the dataset is skewed rather than uniformly distributed. 

### `transcription`

Let's then find some descriptive details about the transcriptions too. Let's begin by getting the distribution of transcription character lengths.

>Retrieve the length of each ``transcription`` in the transcription column of the ``df`` dataset.
>
>Plot an empirical cumulative distribution function (ECDF) of the transcription lengths using Seaborn's ``ecdfplot()`` function.

sns.ecdfplot(df["transcription"].apply(len))

The median length of the transcriptions is around 3000 chracters. This doesn't, however, tell a lot as with NLP single characters are not that telling enough.

## Preprocessing the texts

Here's an outline of the steps you can take to prepare your data for an NLP task. These preprocessing steps are not always necessary or appropriate for every NLP task or dataset. However, they are commonly applied to text data to help improve the accuracy and efficiency of NLP models:

- __Lowercasing the text__: Text data is often case sensitive, meaning that "hello" and "Hello" would be treated as different words by NLP models. Lowercasing the text can help to standardize the text and reduce the dimensionality of the feature space.

- __Removing punctuation__: Punctuation marks such as commas, periods, and question marks do not contribute much to the meaning of text and can add noise to NLP models. Removing punctuation can help to simplify the text and make it easier for models to learn from.

- __Tokenization__: Breaking text data into individual words or subwords (tokens) is a fundamental step in many NLP tasks. Tokenization helps to simplify the text and make it easier to analyze, process, and understand.

- __Removing stop words__: Stop words are common words such as "the", "and", and "a" that do not carry much meaning in text data. Removing stop words can help to simplify the text and reduce the dimensionality of the feature space.

- __Stemming or lemmatization__: Words in text data can have different forms (e.g., "run", "running", "ran"). Stemming or lemmatization is the process of reducing words to their base form, such as "run". This can help to standardize the text and reduce the dimensionality of the feature space.


### Tokenization 

What we're interested in are the __tokens__ of the transcription. In NLP, a token is a sequence of characters that represents a unit of meaning. It is typically a word, but can also be a phrase or a subword. The process of breaking text into tokens is called tokenization, and it is a fundamental step in many NLP tasks, including text classification, machine translation, and sentiment analysis.

Tokenization is necessary because most NLP algorithms and models operate on discrete units of text, rather than on continuous text. By breaking text into tokens, we can analyze, process, and understand the meaning of text more efficiently and effectively.

A token can be defined in many ways, depending on the specific NLP task and the level of granularity required. For example, a token can be a sequence of characters separated by white spaces, a sequence of characters separated by punctuation marks, or a subword obtained through a more complex process such as byte-pair encoding or wordpiece.

Consider the following sentence:

        "a fox was happily jumping in the field with a hedgehog"

To tokenize the text, it can be  split into individual words, resulting in the following list of tokens:

        ["a", "fox", "was", "happily", "jumping", "in", "the", "field", "with", "a", "hedgehog"]


### Removing stop words

Stop word removal is a preprocessing step in NLP that involves removing common words that do not carry much meaning in text data. These common words are known as stop words and typically include words such as "the", "and", "a", and "in".

The goal of stop word removal is to simplify the text and reduce the dimensionality of the feature space, which can improve the efficiency and accuracy of NLP models. Removing stop words can also help to reduce noise in the text and improve the interpretability of the results.

However, stop word removal should be used with care, as removing too many words can lead to loss of information and potential reduction in model performance. Additionally, the list of stop words used may vary depending on the language of the text, the specific NLP task at hand, and the requirements of the downstream NLP model.

For our tokenized sentence, the stop words could very well be `"a"`, `"was"` and `"the"`. After the removal, the tokenized sentence would turn out to be:

        ["fox", "happily", "jumping", "in", "field", "with", "hedgehog"]



### Stemming/Lemmatization

Stemming (or interchangeably lemmatization) is a process in NLP that involves reducing words to their base form, or stem. The goal of stemming is to standardize the text and reduce the dimensionality of the feature space, which can help improve the accuracy and efficiency of NLP models.

There are several algorithms that can be used for stemming, including the Porter stemming algorithm, the Snowball stemming algorithm, and the Lancaster stemming algorithm. These algorithms apply a set of rules to reduce words to their base form, such as removing suffixes or prefixes, and may use language-specific rules or heuristics to handle irregular words.

For example, the Porter stemming algorithm first breaks the sentence into individual words, based on the spaces between the words. In this case, the resulting words would be:

        "fox" would become "fox"
        "happily" would become "happili"
        "jumping" would become "jump"
        "in" would become "in"
        "field" would become "field"
        "with" would become "with"
        "hedgehog" would become "hedgehog"

This would turn the tokens in the initial sentence into:

        ["fox", "happili", "jump", "in", "field", "with", "hedgehog"]

### Data selection

Let's first begin by selecting only the data that we're interested in. Let's select only those medical specialties that have more than 5 % coverage in the whole dataset to have at least somewhat balanced dataset.

>Filter the dataframe to include only the rows corresponding to medical specialties that are represented by more than 5 % of the data, and store the resulting filtered dataframe in a new variable `df_processed`.
>
>Compute and print the frequency of each medical specialty in the filtered dataframe as a proportion of the total number of rows.
>
>Create a horizontal bar chart to visualize the frequency distribution of medical specialties in the filtered dataframe.

specialty_filter = df["medical_specialty"].value_counts(normalize=True) > 0.05
df_processed = df[
    df["medical_specialty"].isin(specialty_filter[specialty_filter].index)
]
print(df_processed["medical_specialty"].value_counts(normalize=True))
df_processed["medical_specialty"].value_counts().plot(
    kind="barh", figsize=(8, 1.5), fontsize=8
)

Let's then select the columns that are of interest for us.

> Select the ``medical_specialty`` and ``transcription`` columns from the processed dataframe

df_processed = df_processed[["medical_specialty", "transcription"]]

# Don't remove this cell
assert df_processed.shape == (2862, 2)

Let's also clean up the medical specialtys. The specialtys seem to have a trailing whitespace. Let's clean the column up a bit. Some specialtys are also multipart. For brevity, let's take only the first part of the specialty. We can always come back to see the whole specialty if need be.

> Split the medical specialty values with `"/"` and `"-"` and select only the first part of the specialty.
>
> Trim trailing whitespaces.

df_processed["medical_specialty"] = df_processed["medical_specialty"].apply(
    lambda x: x.split("/")[0].split("-")[0].strip()
)

# Don't remove this cell
assert any(["Pulmonary" not in c for c in df_processed["medical_specialty"].unique()])

### Preprocessing

Let's perform the preprocessing in one go. The ``preprocess_text`` function is a text preprocessing function in NLP that takes as input a string of text and applies several preprocessing steps in sequence to clean and transform the text for use in downstream NLP tasks.

The preprocessing steps applied in this function include:
1. Lowercasing the text to standardize the text and reduce the dimensionality of the feature space.
1. Removing punctuation from the text to remove noise and simplify the text.
1. Tokenizing the text to break it into individual words, which can be used as features in downstream NLP models.
1. Removing stop words from the text, which are common words that do not carry much meaning in text data, to reduce noise and improve the efficiency and accuracy of downstream NLP models.
1. Stemming/Lemmatizing the words in the text, which involves reducing words to their base form to further reduce the dimensionality of the feature space and improve the accuracy of downstream NLP models.

To accomplish these preprocessing steps, the code utilizes the Natural Language Toolkit (NLTK), a popular library for natural language processing tasks in Python.

NLTK provides a wide range of functionalities and resources for NLP, including various corpora, lexical resources, text processing algorithms, and more. In this code, NLTK is used for the following purposes:

- The ``stopwords`` module from NLTK is used to obtain a set of common stop words in the English language. These stop words are used to filter out words that do not carry significant meaning in the text.
- The ``word_tokenize`` function from NLTK is used to tokenize the text, splitting it into individual words. This function handles the tokenization process by considering punctuation, contractions, and other aspects of natural language.
- The ``WordNetLemmatizer`` class from NLTK is used to lemmatize the words in the text. Lemmatization aims to reduce words to their base or dictionary form (lemmas) to improve the efficiency and accuracy of subsequent NLP tasks.

By leveraging the capabilities of the NLTK library, the ``preprocess_text`` function provides a streamlined way to preprocess text data and prepare it for further analysis or modeling in the field of natural language processing.

# Don't remove this cell
import string
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

nltk.download("stopwords")
nltk.download("punkt")
nltk.download("wordnet")
nltk.download("omw-1.4")


def preprocess_text(text: str) -> str:
    # Lowercase the text
    text = text.lower()

    # Replace punctuation with spaces
    text = re.sub(rf"[{string.punctuation}]", " ", text).strip()

    # Tokenize the text
    words = nltk.word_tokenize(text)

    # Remove stop words
    stop_words = set(stopwords.words("english"))
    words = [word for word in words if word not in stop_words]

    # Lemmatize the words
    lemmatizer = WordNetLemmatizer()
    words = [lemmatizer.lemmatize(word) for word in words]

    # Return the preprocessed text as a string
    return " ".join(words)

# Don't remove this cell
preprocess_text(
    "This is an example,.that kind of looks like, , 1.0,the data in question"
)

Let's then preprocess our data with the function.

>Apply a text preprocessing function to a column of text data in a Pandas dataframe, such as the "transcription" column, and store the preprocessed text in a new column of the same dataframe.
>
>Inspect the first few rows of a Pandas dataframe containing preprocessed text data, to verify that the preprocessing function was applied correctly and that the resulting text data is in the desired format.

df_processed["transcription"] = df_processed["transcription"].apply(preprocess_text)
df_processed.head()

# Don't remove this cell
assert isinstance(df_processed.iloc[0, 1], str)

Let's retrieve the character level lengths of processed transcriptions for comparison.

>Retrieve the length of each ``transcription`` in the transcription column of the ``df_processed`` dataset.
>
>Plot an empirical cumulative distribution function (ECDF) of the transcription lengths using Seaborn's ``ecdfplot()`` function.

sns.ecdfplot(df_processed["transcription"].apply(len))

Median length is now around 2000 characters and the longer transcriptions are also considerably shorter now.

### Splitting the data

The last part is to split the processed data into fine-tuning and validation datasets. 

> Split the processed data into 80 % training (``df_train``) and 20 % validation (`df_validation`).

train_split = 0.8
df_train = df_processed.sample(frac=train_split)
df_validation = df_processed.drop(df_train.index)

# Don't remove this cell
assert 2280 < df_train.shape[0] < 2300
assert 560 < df_validation.shape[0] < 580

### Saving the data

The last step in preprocessing is saving the processed training and validation datasets for later use.

# Don't remove this cell
import os

for data in ["train", "validation"]:
    path = f"../../data/{data}.csv"
    if os.path.isfile(path):
        os.remove(path)

> Persist the training and validation sets with indices as CSVs to the repository's data-folder:
> - Training CSV path: 
>   - Local: `"../../data/train.csv"`
>   - Colab: `"./data/train.csv"`
> - Validation CSV path:
>   - Local: `"../../data/validation.csv"`
>   - Colab: `"./data/validation.csv"`

df_train.to_csv("../../data/train.csv", index=True)
df_validation.to_csv("../../data/validation.csv", index=True)
